requirements.txt
opencv-python
mediapipe
sounddevice
openai-whisper
argparse
torch
